IGAPP-1033: Move icons to common assets folder
