+++ web/src/components/ShelterInformation.tsx
@@ -5,19 +5,19 @@ import styled from 'styled-components'
 
 import { ShelterModel } from 'api-client'
 
-import accessibleIcon from '../assets/shelter/accessible.svg'
-import bathroomIcon from '../assets/shelter/bathroom.svg'
-import bedIcon from '../assets/shelter/bed.svg'
-import calendarIcon from '../assets/shelter/calendar.svg'
-import emailIcon from '../assets/shelter/email.svg'
-import euroIcon from '../assets/shelter/euro.svg'
-import houseIcon from '../assets/shelter/house.svg'
-import keyIcon from '../assets/shelter/key.svg'
-import lgbtqiIcon from '../assets/shelter/lgbtqi.svg'
-import petIcon from '../assets/shelter/pet.svg'
-import phoneIcon from '../assets/shelter/phone.svg'
-import smokingIcon from '../assets/shelter/smoking.svg'
-import timerIcon from '../assets/shelter/timer.svg'
+import accessibleIcon from '../../../assets/icons/accessible.svg'
+import bathroomIcon from '../../../assets/icons/bathroom.svg'
+import bedIcon from '../../../assets/icons/bed.svg'
+import calendarIcon from '../../../assets/icons/calendar.svg'
+import emailIcon from '../../../assets/icons/email.svg'
+import euroIcon from '../../../assets/icons/euro.svg'
+import houseIcon from '../../../assets/icons/house.svg'
+import keyIcon from '../../../assets/icons/key.svg'
+import lgbtqiIcon from '../../../assets/icons/lgbtqi.svg'
+import petIcon from '../../../assets/icons/pet.svg'
+import phoneIcon from '../../../assets/icons/phone.svg'
+import smokingIcon from '../../../assets/icons/smoking.svg'
+import timerIcon from '../../../assets/icons/timer.svg'
 import Caption from './Caption'
 import ShelterContactRequestForm from './ShelterContactRequestForm'
 import ShelterInformationSection from './ShelterInformationSection'
